FW-337. Avoiding using float in timer computation.

+++ openapps/csensors/csensors.h
@@ -22,7 +22,7 @@
 typedef struct {
    coap_resource_desc_t         desc;
    opensensors_resource_desc_t* opensensors_resource;
-   uint32_t                     period;
+   uint16_t                     period;
    opentimer_id_t               timerId;
 } csensors_resource_t;
